bump concourse helm version to 17.0.37 (7.8.3)
* fix asdf-plugin-install.sh to be executed from terragrunt folder
  (same as other scripts by ../scripts/<script-name>

* add using terraform-modules forl local dev (leaving git example)

+++ terragrunt/scritps/asdf-plugin-install.sh
@@ -7,9 +7,15 @@ if [ ! $(which asdf) ]
     exit 1
 fi
 
+if [ ! -s ../.tool-versions ]
+  then
+    echo "ERROR: .tools-versions not found in root directory"
+    exit 1
+fi
+
 echo
 echo ">> Add required plugin as defined in asdf .tool-versions"
-for p in $(cat .tool-versions | awk '{ print $1 }')
+for p in $(cat ../.tool-versions | awk '{ print $1 }')
     do asdf plugin add $p &&  asdf install $p
 done
 echo
